refactoring code and update README
