+++ .gitignore
@@ -1,4 +1,3 @@
 package-lock.json
 /node_modules
-/static/css/style.css.map
 .DS_Store
