+++ README.md
@@ -1,11 +1,11 @@
 # Prerequisite
 
-Installation of npm and Node.js :
+Installation of npm and Node.js.
 
-[Download page](https://nodejs.org/en/download/)
+Follow the installation procedures on this page:
+[download page](https://nodejs.org/en/download/)
 
 To see if you already have Node.js and npm installed and check the installed version, run the following commands:
-
 ```
 node -v
 npm -v
@@ -15,7 +15,6 @@ npm -v
 # Installation
 
 Launch this command to install all dependencies:
-
 ```
 npm install
 ```
@@ -24,13 +23,22 @@ npm install
 
 ## local server
 
+To run the website in a static HTTP server, run this command:
 ```
 npm run dev
 ```
+and go to this url http://127.0.0.1:8080/ to see the website.
+
+
 Press Ctrl-C to stop local server.
+
+***
+
+# Development
 
 ## sass
 
+If you modify scss file, you need to compile to update css file with this command:
 ```
 npm run sass
 ```
@@ -38,6 +46,7 @@ Press Ctrl-C to stop process.
 
 ## eslint
 
+If you modify js file, run this command to check that the code uses the JS standards: 
 ```
 npm run lint
 ```
