+++ public/static/js/main.js
@@ -1,0 +1,8 @@
+import { initDomStructure } from './domStructure.js'
+import { displayFavoriteBooks } from './favorite.js'
+
+// Initialisation de la structure de la page
+initDomStructure()
+
+// Affichage dans les favoris des livres qui sont dans le localStorage
+displayFavoriteBooks()
