+++ sass/layouts/_modalbox.scss
@@ -31,3 +31,4 @@
         }
     }
   }
+
